Move samples manager module to root
- Add `InventoryManager` wrapper on managers
