+++ aurora/interface/menus/__init__.py
@@ -1,7 +1,5 @@
-from .manage_samples import ManageSamplesMenu
 from .visualize_results import CyclingResultsWidget
 
 __all__ = [
-    "ManageSamplesMenu",
     "CyclingResultsWidget",
 ]
