+++ aurora/inventory/__init__.py
@@ -1,0 +1,5 @@
+from .manager import InventoryManager
+
+__all__ = [
+    "InventoryManager",
+]
