+++ aurora/inventory/samples/__init__.py
@@ -1,0 +1,5 @@
+from .manage_samples import ManageSamplesMenu
+
+__all__ = [
+    "ManageSamplesMenu",
+]
