+++ aurora/inventory/samples/manage_samples.py
@@ -4,7 +4,8 @@ import ipywidgets as ipw
 from ipyfilechooser import FileChooser
 
 from aurora.common.models import AvailableSamplesModel, BatteryExperimentModel
-from aurora.interface.sample.sample_filter import SampleFilterWidget
+
+from .sample_filter import SampleFilterWidget
 
 
 class ManageSamplesMenu(ipw.VBox):
